% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_barre.R
\name{gg_barre}
\alias{gg_barre}
\title{Graphique de distribution d'une variable qualitative}
\usage{
gg_barre(df, var, x_lab, type = "stations", couleur = "#D95F02")
}
\arguments{
\item{df}{Nom du dataframe contenant les données.}

\item{var}{Caractère. Nom de la variable qualitative à représenter.}

\item{x_lab}{Caractère. Etiquette de l'axe des x.}

\item{type}{Caractère. Type de dataframe, soit "pêches", soit "stations".}

\item{couleur}{Caractère. Couleur de remplissage des barres.}
}
\value{
Un diagramme ggplot2.
}
\description{
Graphique de distribution d'une variable qualitative
}
