% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_boxplot.R
\name{gg_boxplot}
\alias{gg_boxplot}
\title{Produire un boxplot sur les stations ou les observations}
\usage{
gg_boxplot(
  df,
  var_quant,
  var_quant_lab,
  var_qual,
  var_qual_lab,
  couleur = "#7570B3"
)
}
\arguments{
\item{df}{Dataframe contenant les données.}

\item{var_quant}{Caractère. Nom de la variable quantitative à représenter (ordonnée).}

\item{var_quant_lab}{Caractère. Etiquette de l'axe des y.}

\item{var_qual}{Caractère. Nom de la variable qualitative à représenter (abscisse).}

\item{var_qual_lab}{Caractère. Etiquette de l'axe des x.}

\item{couleur}{Caractère. Couleur de remplissage des barres de l'histogramme.}
}
\value{
Un boxplot ggplot2.
}
\description{
Produire un boxplot sur les stations ou les observations
}
