% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_comp_moy.R
\name{gg_comp_moy}
\alias{gg_comp_moy}
\title{Produire un diagramme barre d'erreur pour comparer les années}
\usage{
gg_comp_moy(df, var, y_lab = NA, couleur = "#7570B3")
}
\arguments{
\item{df}{Caractère. Nom du dataframe contenant les données.}

\item{var}{Caractère. Nom de la variable quantitative à représenter (ordonnée).}

\item{y_lab}{Caractère. Etiquette de l'axe des y.}

\item{couleur}{Caractère. Couleur de remplissage des barres.}
}
\value{
Un errorbar ggplot2.
}
\description{
Produire un diagramme barre d'erreur pour comparer les années
}
