% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_cran_annee.R
\name{gg_cran_annee}
\alias{gg_cran_annee}
\title{Produire un diagramme pour comparer les années sur le cran de réglage et la conductivité}
\usage{
gg_cran_annee(df, var_ordre)
}
\arguments{
\item{df}{Dataframe contenant les données.}

\item{var_ordre}{Caractère. Nom de la variable servant à ordonner les stations en ordonnée.}
}
\value{
Un diagramme ggplot2.
}
\description{
Produire un diagramme pour comparer les années sur le cran de réglage et la conductivité
}
