% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_distri_var_quant.R
\name{gg_distri_var_quant}
\alias{gg_distri_var_quant}
\title{Produire l'histogramme d'une des variables du jdd}
\usage{
gg_distri_var_quant(df, var, xlab, type = "stations", couleur = "#D95F02")
}
\arguments{
\item{df}{Caractère. Nom du dataframe contenant les données.}

\item{var}{Caractère. Nom de la variable à représenter.}

\item{xlab}{Caractère. Etiquette de l'axe des x.}

\item{type}{Caractère. Type de dataframe, soit "pêches", soit "stations".}

\item{couleur}{Caractère. Couleur de remplissage des barres de l'histogramme.}
}
\value{
Un histogramme ggplot.
}
\description{
Produire l'histogramme d'une des variables du jdd
}
