% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_mef_survol.R
\name{gp_mef_survol}
\alias{gp_mef_survol}
\title{Mise en forme des popups ggplotly}
\usage{
gp_mef_survol(g_plotly)
}
\arguments{
\item{g_plotly}{Graphique de classe plotly}
}
\value{
Le graphique dont les popups ont été modifiés.
}
\description{
Mise en forme des popups ggplotly
}
