% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_equation_cran.R
\name{int_equation_cran}
\alias{int_equation_cran}
\title{Produire l'équation mise en forme d'une régression, pour un cran de réglage donné.}
\usage{
int_equation_cran(df, y, x, sel_cran)
}
\arguments{
\item{df}{Dataframe contenant les données.}

\item{y}{Caractère. Nom de la variable quantitative dépendante (ordonnée).}

\item{x}{Caractère. Nom de la variable quantitative explicative (abscisse).}

\item{sel_cran}{Entier. Cran de réglage sélectionné (entre 2 et 4).}
}
\value{
L'équation au format texte.
}
\description{
Produire l'équation mise en forme d'une régression, pour un cran de réglage donné.
}
