% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ordonner_vars.R
\name{mef_ordonner_vars}
\alias{mef_ordonner_vars}
\title{# fonction de mise en forme du df}
\usage{
mef_ordonner_vars(df, ordre)
}
\arguments{
\item{df}{Caractère. Nom du dataframe contenant les données.}

\item{ordre}{Vecteur entier servant à réordonner les colonnes.}
}
\value{
Le dataframe avec les colonnes sélectionnées réordonnées.
}
\description{
# fonction de mise en forme du df
}
