% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bivar_penny.R
\name{model_bivar_penny}
\alias{model_bivar_penny}
\title{Produire un résumé mis en forme du modèle penny_15_centre ~ variable}
\usage{
model_bivar_penny(df, var)
}
\arguments{
\item{df}{Dataframe contenant les données.}

\item{var}{Caractère. Nom de la variable explicative.}
}
\value{
Une liste contenant les dirrérents éléments du modèle (R2, p-value, etc.).
}
\description{
Produire un résumé mis en forme du modèle penny_15_centre ~ variable
}
\examples{
#' @examples \dontrun{
model_bivar_penny(df = prov, var = "intensite")
}
}
