% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_cor.R
\name{g_cor}
\alias{g_cor}
\title{Graphique de matrice de corrélation}
\usage{
g_cor(df, order = "original")
}
\arguments{
\item{df}{Dataframe contenant uniquement les variables à corréler.}

\item{order}{Caractère. Ordre des variables sur le graphique qui est soit celui
du dataframe ("original") soit fonction des proximités en CAH ("hclust").}
}
\value{
La graphique.
}
\description{
Graphique de matrice de corrélation
}
