% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_histo.R
\name{gg_histo}
\alias{gg_histo}
\title{Produire l'histogramme d'une des variables du jdd}
\usage{
gg_histo(df, var, x_lab, type = "stations", couleur = "#D95F02")
}
\arguments{
\item{df}{Nom du dataframe contenant les données.}

\item{var}{Nom de la variable à représenter.}

\item{x_lab}{Caractère. Etiquette de l'axe des x.}

\item{type}{Caractère. Type de dataframe, soit "pêches", soit "stations".}

\item{couleur}{Caractère. Couleur de remplissage des barres de l'histogramme.}
}
\value{
Un histogramme ggplot.
}
\description{
Produire l'histogramme d'une des variables du jdd
}
